[GUPPY-1751] Use fixed tensor sizes for benchmarking
Our Windows CI runner uses a 1080 Ti which takes ~15s in the cuBLAS
benchmark when lengths are (2048,512,1024), and that's unreasonable
startup time so we're forced to use smaller lengths.

This way it's always the same fixed overhead no matter what the first
call to matmul_f16() is (ie larger tensors would have taken longer to
benchmark).