[cli] options to set aligner/writer threads
Previously we had only one option to set number of
threads, which set the thread count for both BAM writing and
alignment. However, this led to 2x the number of threads being
created since the same thread count was used for writing and alignment.

This MR adds new options to explicitly set the thread count for
each, and if not specified, allocates threads automatically up
to the max number of threads possible.