[aligner] fix alignment count differences w/ mm2
What?
The number of primary/secondary/supplementary
alignments being returned by dorado aligner was not
matching the counts from mm2. This commit fixes that.

How?
Pass the qname string to the mm_map() call because
the minimap2 equivalent call does. And passing the qname
string generates a hash internally which affects the
logic within mm2 chaining.