[data_loader] cache file handles
update code to cache the file handles for the POD5 files.
this improves performance and gets rid of bubbles in the GPU
utilization during duplex.
caveat is that by default systems have a limit of 1024 file
handles open at a time. this can be changed easily in case there
are more than 1024 POD5 files to process.