[bam_utils] write ubam when output is a pipe
Instead of writing BAM write uncompressed BAM when
output is a pipe to another process. This saves both compression
and decompression time and frees up cpu cycles.

Note that this is currently only available for unix platforms
but not on windows. On Windows the output is still BAM if it's
piped into another process.