Revert parent_read_id handling in SAM files to refine design later
This commit removes the addition of the 'pi' tag to the output SAM file
for representing parent read IDs. The decision to revert this change is
due to the need for further discussion and refinement of the design details
regarding how parent read IDs should be represented in SAM files. This will
ensure a more thoughtful implementation that does not rush the design process.