[INSTX-459] Don't feed the output of malloc() into a unique_ptr
unique_ptr<> by default will assume that the data it's holding will
have come from new() rather than malloc(), and so when it gets delete'd
we have a mismatch since the memory may not come from the same heap.

Instead use a std::array<> since the data we're copying is fixed size
and this avoids us having to go to the heap for each 16 byte object.