[duplex] make gpu mutex optional
the GPU mutex was causing some bubbles in the duplex
pipeline whensimplex was faster than duplex. removing that mutex
helps. but we kept it optional in case users run into weird
OOM cases.

also adjust the conditional for basecaller input queue to wait
if there are too many working reads in flight.