[duplex] remove bubbles in duplex with gpu mutex
the bubbles in the pipeline were caused by the lock
that protected the unordered_set which stored the
mutexes per gpu index. this was addressed by declaring
an array of mutexes > number of gpus we expect to see
in practice.

because we can now always use the gpu mutexes, the
--guard-gpus options is no longer needed in the duplex
pipeline.

Additionally, remove a device synchronize which was showing
up in the nvtx trace but seemed unnecessary.