[INSTX-765] Make sure that tests can see the libs they need
This was Just Work-ing on CI for unknown reasons (likely hts-3.dll is
installed somewhere), but with the new static torch lib there's more
that are missing. Pointing the tests at the install location lets them
find the libs they need.

This also fixes an issue where using a custom install prefix caused the
tests to fail to run on Windows.