[qscore] exclude some bases from qscore calculation
The mean q-score being calculated for some short reads is
artificially low becuase of some bases at the beginning
of the read having a low q-score. This change allows
setting a start position for the mean-qscore calculation
on a per model basis (default start pos is 0, i.e. no
bases are skipped).

Closes DOR-197