[ReadSplit] Fix read ids for non-split reads
A bug was introduced wherein reads which were not split
were still generating new read ids. However, because
not all of the state of the reads were set for unsplit read
ids, it was causing the same read id to be generated for every
one of them. This caused downstream issues with logging and
created incorrect outputs.