[polyA] improve signal range detection
Improve the signal range detection algorithm
This new method is more brute force and simplified
compared to previous ones. It uses a fixed variance
threshold and simply walks the signal around a
suggested anchor and finds the longest range of a
"flat" signal.
Also relax the constraint for primer edit distance
score.