[DOR-215] Refactor the BasecallerNode to have a single ownership of the Read object
The chunk info doesn't need to be part of the Read since it's not used
by any other nodes, so split that out into its own thing. That then
allows us to keep the read in one place and only move it on once all of
its chunks are processed.

stitch_chunks() could also be refactored to drop its dependency on Read
but that's not required right now so can be done another day.