[DOR-215] Add ReadPtr as a smart pointer to hold Reads
This smart pointer is a lot like a unique_ptr in that there's a single
mutable owner of the Read, however it can also hand out immutable views
that have a strong ownership of the Read too. This allows us to hand
off parts of the Read to processing that doesn't modify the Read, while
not having to worry about the lifetime of the Read since all views have
strong ownership.