The iOS build of htslib was non-functional in several ways.
Cross compilation flags were incomplete, leading to a MacOS build.  This adds the requisite flags.

The build occurred in the source tree, and htslib's autotools build appears to have a bug that leaves residual incompatible object files, that led to linker errors.  This moves the build directory so that it is target-specific.

The install name modification required for iOS devices to find libhts.dylib wasn't being done automatically.  Now it's invoked via ExternalProject_Add.