[DOR-215] Remove TODO that wasn't an issue
The lifetime of the |Read| is fine since we use |m_reads_in_flight_ctr|
to keep track of when to move the read on to the next node.

However we do need to make |is_duplex_parent| atomic since it can be
assigned by multiple nodes at the same time.