Add modbase chunks to queue after setting read in working_read
Chunks can be processed as soon as they are pushed to the `chunk_queue`.
Part of the processing involves reading from working_read->read->seq.
If the `input_worker_thread` had not progressed sufficiently to set this
before it was required, we would dereference a nullptr.
Queue the chunks locally and only pass them on once we're ready.