Attempt to get tests that use CUDA working when ASAN is enabled
We already do this in ont_core_cpp since it was more of an issue when
talking to the CUDA API, but since torch talks to CUDA for us it gets
to decide how to handle the issue and it looks like it just prints a
warning and continues on assuming no CUDA support. This is fine for end
users, but for automated testing it means we were skipping tests
without realising.

See https://github.com/google/sanitizers/issues/629 for more details.