Only use a PCH if ccache isn't available
The PCH generation is useful for developers that don't have ccache, but
for developers that do it's a slowdown since the PCH is generated into
the build folder. In addition to this, simply adding or removing a file
from the |dorado_utils| target causes the PCH to be re-generated and
hence everything has to recompile.

Instead, let's rely on the fact that most files don't change often and
hence can be provided by ccache for developers that have it, and those
that don't can fall back to the benefit of using a PCH over a plain
build.

This reduces the time to compile a fresh macOS build from ~3mins to
18s.