Use DORADO_GPU_BUILD rather than !defined(__x86_64__)
Developers on an Apple Silicon machine without Xcode installed should
be able to disable DORADO_GPU_BUILD to get the build working with the
same features that an Intel build would have. This was the only case
where we assumed !x86==>GPU, so let's fix that up to unblock those
developers.