[DOR-347] Add custom cert location handling
We link to a static Ubuntu OpenSSL so other distributions (Centos7)
fail to download models because they can't verify the server's
identity. Parsing /etc/os-release tells us which distro we're on and we
can make an educated guess at where they are, but the environment
variable dorado_cert_file_path has also been added so that users can
provide the correct path if we guess wrong.