[INSTX-2496] Fix a crash with GCC8 seen on CI
The following crash was seen on CI when compiling with GCC 8.4.0 and
with ASAN enabled:
  ==11805==ERROR: AddressSanitizer: SEGV on unknown address 0x001200000000 (pc 0xffff68c2296c bp 0xffffc1c19400 sp 0xffffc1c19400 T0)
  ==11805==The signal is caused by a WRITE memory access.
    #0 0xffff68c2296b  (/lib/aarch64-linux-gnu/libc.so.6+0x8396b)
    #1 0xffff876bedc3  (/usr/lib/aarch64-linux-gnu/libasan.so.5+0x36dc3)
    #2 0xaaaad510039b in _GLOBAL__sub_I_00099_1__ZN6dorado8splitter7subreadERKNS_11SimplexReadESt8optionalISt4pairImmEES6_ (dorado_tests+0x17639b)
    #3 0xaaaad5c9ea1f in __libc_csu_init (dorado_tests+0xd14a1f)

The static initialiser that calls into libasan is one of ASANs inserted
init functions (__asan_init, __asan_version_mismatch_check_v8, or
__asan_register_globals) though gdb refuses to ptrace so I can't tell
which. None of them should crash so this feels like a compiler bug,
though I'm unable to determine exactly what's triggering it since a
disassembly of the static initialiser doesn't change between the broken
and "fixed" versions of the code (plt addresses notwithstanding).