Correctly trim modbase tags for reverse strand alignments
This requires the trim interval derived from the adapter
or barcodes be reverse complemented before applying to
modbase tag trimming because the modbase tags are all relative to
the original sequence that was basecalled.

Also add a check before writing BAM records that ensure the sequence
length is the same as the MN tag (if it's available).