[demux] handle input file in current directory
this was previously failing because the parent directory
for a file in the current directory is an empty string, which
was later being used to determine an absolute path. but absolute
path for an empty string cannot be determined, so filesystem
API was throwing an error. this is now fixed so the absolute
path of the input file is used.