[resume] Fix resume loading for split reads
Previously the resume loader simply loaded the
read ids from the records in the input BAM. But some
of those records were from split reads, which wouldn't
match a read id in the input dataset. Instead, for split
reads, the parent read id needs to be extracted. This change
fixes that bug.

Closes DOR-599