[DOR-610] Remove escape code on Linux too
\033[K erases to the end of the line and for some reason breaks the
progress bar for the aligner and demux frontends. The only thing I can
see that helping with is when the estimated time shrinks, but that
would only leave an additional ] at the end of the line which isn't too
bad. The commit adding it also doesn't state why it was required, and
testing the other frontends shows it continuing to work as expected.