Fix test_archive:linux:arm64:20.04 jobs
The glob in `tar -xzf archive/dorado-*.gz` is expanded by the shell so
if there's more than one match we end up asking `tar` to extract an
archive from another archive, which is rightly complains doesn't exist.
Instead do this step inside a folder that we can nuke every time it's
run. Note that this is only impacting test_archive:linux:arm64:20.04
since the runner that runs that job doesn't do any builds and hence
gitlab never runs a git clean on the repo, leading to stale files.