Allow dependent targets to be built in parallel when using Makefiles
Linking can take a while in CI and we spend a lot of time waiting for
dependencies to link before starting to build the next library.
Since most of our targets have no side effects we can enable this flag
and cmake will realise that there's no side effects and build those
that it can in parallel.

This saves ~20% total time on a Linux build (7m41s -> 6m15s).