Don't enable ccache for sanitizer builds
From testing, a bad ccache can be up to 2x slower than using a PCH vs a
good cache which can reduce it 5x. We don't seem to be getting anywhere
near that improvement in CI, so disable it. This may help the other
builds too since it's a per-runner cache and hence shared between all
jobs on a runner.