[DOR-661] Enable symbols a different way for the valgrind build
Using a Debug build should have worked too, but valgrind seems to think
that something in torch's thread-local code is misbehaving if we use a
Debug build. The issue goes away in a Release build which is odd
because torch is a pre-built library that we link to. All other issues
repro in both builds, so using a Release build shouldn't be an issue.