[DOR-777] Enforce an alignment on the size of the time slices in MHA
When the metal debug layer is enabled it complains about a buffer index
not starting at an offset divisible by 4. This also happens if we use
our naive implementation where the call to |mask.logical_not()| dies to
the same issue. Disabling the mask via |mha_use_mask| works around the
issue, however that's obviously not a valid fix.