[INSTX-5832] Embed kernel source into the compiled library
This allows Xcode to show where in the kernels we're spending out time,
and rather than everyone having to rediscover it let's keep it enabled
so that it Just Works in the future.

Also add -O2 since both -O2 and -ffast-math are the defaults but if
we're adding one then we may as well add the other.