[INSTX-6929] Handle the case where we're passed an empty folder
When given an empty folder we'll end up finding no files to process, so
skip doing anything. Note that this isn't an error since we didn't fail
to process any of the files that we found.

This code is duplicated between the 2 frontends and only really affects
demux, but best to be consistent.