Package the tests as a single file when moving them between jobs
Gitlab doesn't appear to remove old folders when pulling down artifacts
with the same path, which means that the cram tests from other branches
can end up in the tests to run. Since those tests are looking at new or
changed functionality we get test failures.

Instead package the tests into a single file so that it clobbers
whatever happens to be on the runner.