INSTX-9750: Increase `max_split_size_mb` for Orin.
For some reason (torch bug?) the allocator isn't correctly reusing allocations in the tx model on Orin when this value is small. This leads to it continually allocating new memory blocks until eventually it grinds to a halt (since Orin uses unified memory, this is also our RAM).

We should test reducing this again once we're on torch 2.7 in case they've fixed it.