[INSTX-11681] Drop wonky optimisation
Unlike the BasecallerNode and ModBaseChunkCallerNode were we process an
element at a time, here we erase the whole batch in a loop so we don't
benefit from trying to delay destructors (unless we allocated another
buffer to move them to). I don't think this change had much of, if any,
impact so revert it.