[INSTX-11989] Add support for saving the motif hits
The ont_core client spends a lot of its time recomputing this, but
we've already done the hard work on the server so hand it off to the
client instead of making it do it again.

Note that only the simplex hits are kept since we don't support duplex
modbase in ont_core.